+++ GestionTiendaRafael/src/dao/interfaces/ClienteDAO.java
@@ -7,8 +7,8 @@ import modelo.entidades.Cliente;
 public interface ClienteDAO {
     boolean agregarCliente(Cliente cliente);
     boolean actualizarCliente(Cliente cliente);
-    boolean eliminarCliente(long idCliente);
-    Cliente obtenerClientePorId(long idCliente);
+    boolean eliminarCliente(int idCliente);
+    Cliente obtenerClientePorId(int idCliente);
     List<Cliente> obtenerTodosLosClientes();
     List<Cliente> obtenerClientesPorTipo(String tipoCliente);
 }
